+++ SeniorProjectW/src/Line.java
@@ -5,7 +5,7 @@ public class Line {
 String descrip;
 String lineName;
 String lineAcro;
-float lineLength;
+int lineLength;
 String lineType;
 String direction;
 String start;
@@ -14,7 +14,7 @@ List<Object> Stations;
 int numberOfStation;
 int Linecapacity;
 
-	public Line (String name, String acro, String type, String dir, float length, String strSta, String endSta, int numSta, int capSta, String descr ) {
+	public Line (String name, String acro, String type, String dir, int length, String strSta, String endSta, int numSta, int capSta, String descr ) {
 		 descrip=descr;
 		 lineName=name;
 		 lineAcro=acro;
